Tabulate cases with condition in causal chain by age and site of acquisition

Tabulate cases with condition in causal chain by age and site of acquisition

calc_cc_allcases_by_age_acq(d, condition, age_groups = NULL, sites = NULL)

Arguments

d
A data object obtained from load_data.

condition
A string specifying the condition to tabulate causal chain presence for. To list all possibilities, see valid_conditions.

age_groups
Optional vector of age groups. See valid_age_groups.

sites
An optional vector of site names to include in the tabulation. If not provided, all sites will be used. See valid_sites.

Examples

calc_cc_allcases_by_age_acq(mock,
  condition = "Klebsiella pneumoniae",
  age_groups = valid_age_subcats(mock)
)

#> $version
#> # A tibble: 1 x 3
#>   dataset_name       dataset_version dataset_release_date
#>   <chr>                        <dbl> <date>              
#> 1 CHAMPS Level2 Data             4.1 2020-08-01          
#> 
#> $table
#> # A tibble: 7 x 4
#> # Groups:   age_group_subcat [7]
#>   age_group_subcat                         Community Facility Total
#>   <fct>                                        <int>    <int> <int>
#> 1 Death in the first 24 hours                      5        0     5
#> 2 Early Neonate (24-72 hours)                      4        6    10
#> 3 Early Neonate (72+hrs to 6 days)                 1       14    15
#> 4 Late Neonate (7 to 27 days)                      2       12    14
#> 5 Infant (28 days to less than 6 months)           1        1     2
#> 6 Infant (6 months to less than 12 months)         6        7    13
#> 7 Child (12 months to less than 60 Months)        13        8    21
#> 
#> $denominators
#> # A tibble: 9 x 2
#>   age_group_subcat                             n
#>   <fct>                                    <int>
#> 1 Stillbirth                                 164
#> 2 Death in the first 24 hours                 93
#> 3 Early Neonate (24-72 hours)                 43
#> 4 Early Neonate (72+hrs to 6 days)            24
#> 5 Late Neonate (7 to 27 days)                 36
#> 6 Infant (28 days to less than 6 months)      15
#> 7 Infant (6 months to less than 12 months)    46
#> 8 Child (12 months to less than 60 Months)    70
#> 9 NA                                           4
#> 
#> attr(,"class")
#> [1] "champs_computed"             "calc_cc_allcases_by_age_acq"