Tabulate the number of cases where the pathogen(s) are detected in TAC results in the context of all cases by site and by case type

Tabulate the number of cases where the pathogen(s) are detected in TAC results in the context of all cases by site and by case type

calc_detected_allcases_by_site_casetype(
  d,
  condition,
  pathogen,
  tac_variable = NULL,
  sites = NULL,
  specimen_types = NULL
)

Arguments

d
A data object obtained from load_data.

condition
A string specifying the condition to tabulate causal chain presence for. To list all possibilities, see valid_conditions.

pathogen
A string specifying the pathogen to count positive cases of in the TAC results.

tac_variable
The variable to check for in the TAC laboratory table to count as having completed TAC (currently not used).

sites
An optional vector of site names to include in the tabulation. If not provided, all sites will be used. See valid_sites.

specimen_types
An optional vector of specimen types to include in the calculation. If not provided, all specimen types will be used. See valid_specimen_types.

Examples

if (FALSE) {
d <- load_data(...)
specimen_types <- c(
  "Cerebrospinal fluid sample",
  "Tissue specimen from lung",
  "Whole blood",
  "Rectal swab",
  "Plasma or spun blood specimen"
)

calc_detected_allcases_by_site_casetype(d,
  condition = "GBS",
  pathogen = "Group B Streptococcus",
  tac_variable = "bld_gbst_1",
  specimen_types = specimen_types)
}