Tabulate the number of cases where the pathogen(s) are detected in TAC results by DeCoDe result (in causal chain or not, etc.) and by either site or age

Tabulate the number of cases where the pathogen(s) are detected in TAC results by DeCoDe result (in causal chain or not, etc.) and by either site or age

calc_detected_by_decode(
  d,
  by = "site",
  condition,
  pathogen,
  icds,
  sites = NULL,
  specimen_types = NULL
)

Arguments

d
A data object obtained from load_data.

by
One of either "site" or "age", indicating the second dimension of tabulation.

condition
A string specifying the condition to tabulate causal chain presence for. To list all possibilities, see valid_conditions.

pathogen
A string specifying the pathogen to count positive cases of in the TAC results.

icds
A vector of ICD10 codes to check for the DeCoDe result of "Contributing (P2)".

sites
An optional vector of site names to include in the tabulation. If not provided, all sites will be used. See valid_sites.

specimen_types
An optional vector of specimen types to include in the calculation. If not provided, all specimen types will be used. See valid_specimen_types.

Examples

specimen_types <- c(
  "Cerebrospinal fluid sample",
  "Tissue specimen from lung",
  "Whole blood",
  "Rectal swab",
  "Plasma or spun blood specimen"
)

calc_detected_by_decode(mock,
  by = "site",
  condition = "Streptococcus agalactiae",
  pathogen = "Group B Streptococcus",
  icds = c("P36.0", "A40.1", "P23.3", "G00.2"),
  specimen_types = specimen_types)

#> $version
#> # A tibble: 1 x 3
#>   dataset_name       dataset_version dataset_release_date
#>   <chr>                        <dbl> <date>              
#> 1 CHAMPS Level2 Data             4.1 2020-08-01          
#> 
#> $condition
#> [1] "Streptococcus agalactiae"
#> 
#> $pathogen
#> [1] "Group B Streptococcus"
#> 
#> $icds
#> [1] "P36.0" "A40.1" "P23.3" "G00.2"
#> 
#> $sites
#> NULL
#> 
#> $specimen_types
#> [1] "Cerebrospinal fluid sample"    "Tissue specimen from lung"    
#> [3] "Whole blood"                   "Rectal swab"                  
#> [5] "Plasma or spun blood specimen"
#> 
#> $numerator
#>                    nn$site
#> nn$result           Fipad Jbqcm Ljfjy Nkhzh Wutoi Xxyhs Yqatt Sum
#>   In causal chain       0     0     2     2     1     1     2   8
#>   Contributing (P2)     0     1     0     0     0     0     0   1
#>   Other Infectious      1     4     4     4     3     4     0  20
#>   Not in CC/no ID       1     1     0     0     1     0     0   3
#>   Sum                   2     6     6     6     5     5     2  32
#> 
#> attr(,"class")
#> [1] "champs_computed"    "detected_by_decode"