Tabulate TAC top pathogens detected and causal chain

Tabulate TAC top pathogens detected and causal chain

calc_top_tac_pathogens_cc(
  d,
  condition,
  specimen_types = NULL,
  specimen_abbrv = NULL,
  age_groups = NULL,
  sites = NULL
)

Arguments

d
A data object obtained from load_data.

condition
A string specifying the condition to tabulate causal chain presence for. To list all possibilities, see valid_conditions.

specimen_types
An optional vector of specimen types to include as columns in the table.

specimen_abbrv
An optional vector of abbreviations to use for the specified specimen_types. calculation. If not provided, all specimen types will be used. See valid_specimen_types.

age_groups
Optional vector of age groups. See valid_age_groups.

sites
An optional vector of site names to include in the tabulation. If not provided, all sites will be used. See valid_sites.

Examples

calc_top_tac_pathogens_cc(mock,
  condition = "Lower respiratory infections",
  age_groups = c(
    "Infant (28 days to less than 12 months)",
    "Child (12 months to less than 60 Months)"),
  specimen_types = c(
    "Nasopharyngeal and Oropharyngeal swab",
    "Tissue specimen from lung"),
  specimen_abbrv = c("# NP+", "# Lung+")
)

#> $version
#> # A tibble: 1 x 3
#>   dataset_name       dataset_version dataset_release_date
#>   <chr>                        <dbl> <date>              
#> 1 CHAMPS Level2 Data             4.1 2020-08-01          
#> 
#> $table
#> # A tibble: 36 x 5
#>    pathogen   `# NP+ (n=135)` `# Lung+ (n=135… `# in CC (n=135… `# in CC as LRI…
#>    <chr>                <int>            <int>            <int>            <int>
#>  1 Klebsiell…              92               57               36               19
#>  2 Cytomegal…              85               51               10                6
#>  3 Streptoco…              84               48               28               24
#>  4 Haemophil…              81               49               12                7
#>  5 Staphyloc…              73               27                8                6
#>  6 Moraxella…              71               38                1               NA
#>  7 Rhinovirus              60               22                2                2
#>  8 Adenovirus              33               10                5               NA
#>  9 Acinetoba…              31                9                3               NA
#> 10 Pseudomon…              31               12                4                2
#> # … with 26 more rows
#> 
#> $n
#> [1] 135
#> 
#> $n_condition
#> [1] 77
#> 
#> attr(,"class")
#> [1] "champs_computed"           "calc_top_tac_pathogens_cc"